\name{engel}
\alias{engel}

\title{
  engel
}
\description{
  Engel food expenditure data used in Koenker and Bassett(1982). This is a regression data set consisting of 235 observations on income and expenditure on food for Belgian working class households.
}
\usage{
  data(engel)
}
\format{
  A data frame containing 235 observations on 2 variables
  \describe{
    \item{\code{income }}{annual household income in Belgian francs}
    \item{\code{foodexp}}{annual household food expenditure in Belgian francs}
  }
}

\keyword{datasets}
