% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDLM.R
\name{GDLM}
\alias{GDLM}
\title{Linear regression using R.}
\usage{
GDLM(x, y, crit = 1e-08, maxiter, Intercept = T)
}
\arguments{
\item{x}{The observed value of independent variables x should be a vector or matrix.(numeric)}

\item{y}{The observed value of dependent variable and y should be a vector. (numeric)}

\item{crit}{The value that controls convergence. (numeric)}

\item{maxiter}{Maximum number of iteration steps. (numeric)}

\item{Intercept}{if TRUE then the reslut including the intercept coefficient (logical)}
}
\value{
the regression coefficients
}
\description{
Linear regression  by gradient descent using R.
}
\examples{
\dontrun{
x1=rnorm(100);x2=rnorm(100)
x=cbind(x1,x2)
theta=c(1.1,2)
y=x\%*\%theta+3+rnorm(100)
GDLM(x,y,crit=1e-12,500,Intercept=T)
}
}
