% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrlinearp.R
\name{qrlinear}
\alias{qrlinear}
\title{Quantile regression using R.}
\usage{
qrlinear(x, y, tau, Intercept = T)
}
\arguments{
\item{x}{the observed value of independent variables x should be a vector or matrix.(numeric)}

\item{y}{the observed value of dependent variable and y should be a vector. (numeric)}

\item{tau}{the quantile level and tau should be in (0,1). (numeric)}

\item{Intercept}{if TRUE then the reslut including the intercept coefficient (logical)}
}
\value{
the regression coefficients
}
\description{
Quantile regression by Linear Programming using R. Which means turning Quantile regression into  a linear programming problem
}
\examples{
\dontrun{
data(engel)
b1hat<-matrix(NA,ncol=2,nrow=6);b2hat<-matrix(NA,ncol=2,nrow=6)
taus <- c(.05,.1,.25,.75,.9,.95)
for(i in 1:length(taus)){
 b1hat[i,]=qrlinear(income,foodexp,taus[i],Intercept=T)
 b2hat[i,]=rq(foodexp~income,taus[i])$coef
 }
 print(b1hat);print(b2hat)
}
}
