% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdCaucy.R
\name{rdCauchy}
\alias{rdCauchy}
\title{generate random variables from a Cauchy distribution using R.}
\usage{
rdCauchy(n, sigma, x0, u, lamada)
}
\arguments{
\item{n}{length of the Markov chain(numeric)}

\item{sigma}{standard error of the proposal distribution (numeric)}

\item{x0}{initial value (numeric)}

\item{u}{parameter of cauchy distribution. (numeric)}

\item{lamada}{parameter of cauchy distribution. (numeric)}
}
\value{
the Markov chain and the number of refuse.
}
\description{
Use the Metropolis-Hastings sampler to generate random variables from a Cauchy distribution.
}
\examples{
\dontrun{
n<-10000
cauchy<-rdCauchy(n,sigma=0.5,x0=0,u=0,lamada=1)
refuse.pr<-cauchy$k/n
refuse.pr
hist(cauchy$x[1001:n],freq=F,main="density of cauchy",breaks=60)
curve(dcauchy(x,0,1),add=TRUE)
}
}
