\name{StatComp18026-package}
\alias{StatComp18026-package}
\docType{package}
\title{
  This package provides several functions for benchmarking R.
}
\description{
  R version of gibbs sampler and vaac function are provided in this package (\code{gibbsR} and \code{vaccR} )
}
\author{
Tao Xu

Maintainer: Tao Xu <xu0130@mail.ustc.edu.cn>
}
\keyword{ package }
